requirements.txt
discord.py==2.0.1
urllib3==1.26.5
